fix imports in api index and rename hooks to use-api
