+++ src/utils/api/index.ts
@@ -1,2 +1,2 @@
-export * from '@/hooks/hooks'
+export * from '@/hooks/use-api'
 export * from '@/utils/api/metacom'
